+++ src/main/java/nextstep/auth/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.auth.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
